Move BaseDbRunner from postgres to separate folder
